fix: resolve all TypeScript type-checking errors across 25 templates
Fix three categories of TypeScript errors introduced by the money@0.3.0
IMonetaryAmount migration:

1. CurrencyCode enum (13 templates): Change monetary() helper signatures
   from `currencyCode: string` to `currencyCode: CurrencyCode` since
   callers already pass typed enum values.

2. IMonetaryAmount arithmetic (6 templates): Extract .doubleValue for
   numeric operations; construct proper IMonetaryAmount objects for
   return values instead of returning plain numbers.

3. Model/generated fixes (3 templates):
   - fragile-goods: remove duplicate MonetaryAmount concept from model
   - perishable-goods: use Double fields where logic treats values as scalars
   - latedeliveryandpenalty-currency-conversion: rename local
     CurrencyConversion to ExchangeRate to avoid collision with
     money@0.3.0.CurrencyConversion

Remove fragile-goods from expectedFailures (now passes render test).
All 152 non-expected-failure tests pass; typeCheckingFailures set is empty.

+++ src/fixed-interests/logic/logic.ts
@@ -32,11 +32,11 @@ function monthlyPaymentFormula(loanAmount: number, rate: number, loanDuration: n
 // @ts-ignore TemplateLogic is imported by the runtime
 class FixedInterestsLogic extends TemplateLogic<ITemplateModel> {
     async trigger(data: ITemplateModel, request: IFixedInterestsRequest): Promise<FixedInterestsResponse> {
-        const monthly = monthlyPaymentFormula(data.loanAmount, data.rate, data.loanDuration);
+        const monthly = monthlyPaymentFormula(data.loanAmount.doubleValue, data.rate, data.loanDuration);
 
         return {
             result: {
-                output: `loan for the amount of ${data.loanAmount}`,
+                output: `loan for the amount of ${data.loanAmount.doubleValue}`,
                 $timestamp: new Date(),
                 $class: 'org.accordproject.fixedinterests@0.2.0.FixedInterestsResponse'
             }
